+++ titiler/main.py
@@ -3,18 +3,17 @@
 from . import settings, version
 from .endpoints import cog, mosaic, stac, tms
 from .errors import DEFAULT_STATUS_CODES, add_exception_handlers
+from .middleware import CacheControlMiddleware, TotalTimeMiddleware
 
 from fastapi import FastAPI
 
 from starlette.middleware.cors import CORSMiddleware
 from starlette.middleware.gzip import GZipMiddleware
-from starlette.requests import Request
 
 api_settings = settings.ApiSettings()
 
 app = FastAPI(
     title=api_settings.name,
-    openapi_url="/api/v1/openapi.json",
     description="A lightweight Cloud Optimized GeoTIFF tile server",
     version=version,
 )
@@ -37,20 +36,8 @@ if api_settings.cors_origins:
     )
 
 app.add_middleware(GZipMiddleware, minimum_size=0)
-
-
-@app.middleware("http")
-async def header_middleware(request: Request, call_next):
-    """Add custom header."""
-    response = await call_next(request)
-    if (
-        not response.headers.get("Cache-Control")
-        and api_settings.cachecontrol
-        and request.method in ["HEAD", "GET"]
-        and response.status_code < 500
-    ):
-        response.headers["Cache-Control"] = api_settings.cachecontrol
-    return response
+app.add_middleware(CacheControlMiddleware, cachecontrol=api_settings.cachecontrol)
+app.add_middleware(TotalTimeMiddleware)
 
 
 @app.get("/ping", description="Health Check", tags=["Health Check"])