+++ titiler/middleware.py
@@ -1,0 +1,45 @@
+"""Titiler middlewares."""
+
+import time
+from typing import Optional
+
+from starlette.middleware.base import BaseHTTPMiddleware
+from starlette.requests import Request
+from starlette.types import ASGIApp
+
+
+class CacheControlMiddleware(BaseHTTPMiddleware):
+    """MiddleWare to add CacheControl in response headers."""
+
+    def __init__(self, app: ASGIApp, cachecontrol: Optional[str] = None) -> None:
+        """Init Middleware."""
+        super().__init__(app)
+        self.cachecontrol = cachecontrol
+
+    async def dispatch(self, request: Request, call_next):
+        """Add cache-control."""
+        response = await call_next(request)
+        if (
+            not response.headers.get("Cache-Control")
+            and self.cachecontrol
+            and request.method in ["HEAD", "GET"]
+            and response.status_code < 500
+        ):
+            response.headers["Cache-Control"] = self.cachecontrol
+        return response
+
+
+class TotalTimeMiddleware(BaseHTTPMiddleware):
+    """MiddleWare to add Total process time in response headers."""
+
+    async def dispatch(self, request: Request, call_next):
+        """Add X-Process-Time."""
+        start_time = time.time()
+        response = await call_next(request)
+        process_time = time.time() - start_time
+        timings = response.headers.get("Server-Timing")
+        app_time = "total;dur={}".format(round(process_time * 1000, 2))
+        response.headers["Server-Timing"] = (
+            f"{timings}, {app_time}" if timings else app_time
+        )
+        return response